Add base HoloLens project
Forgot to init git and after I completely crashed the editor, I have realized it is time to do so. This commit has all the base
implementations and will be developed further. The only thing missing is the fragment output in CtShader code. That one
apparently caused the issue of editor crashing. I have yet to investigate further to find out why it suddenly became a problem.