Add asset names and histogram code
Histogram code uses compute shader but with an atomic add, which is better than CPU but
still there is a lot to be desired. Atomics are really bad on gpu with that many threads, it would
be better to use some buffers per thread and then reducing the counters combined
with some sync blocks or something. But for now it will do.