Add transfer function drawer and controller
This is TF drawer for 1d, 2d and Ellipses. It was ported from the regular unity prototype project and unfortunately done a bit quickly and under pressure.
Additionally there were some fallse presumptions, which draw me into a trap that I did not get out. All of this resulted in a very unstable code, which
should be refactored, when there is more time. First thing to do is seperate a few of the slider controllers to a new unity object, after that
the class needs to be stripped of opacityPoints. The base class should only deal with one collection of points and the the user should set it up correctly
in the editor.