Add support for the image creation and filtering on the gpu
So basically this commits adds the support to filter and create an image on the gpu (normalize). It is probably not the best, but if you do
not need to store the assets (create asset function), it basically takes a few seconds to process (compared to CPU with 3 minutes).
This is for data size around 256 x 256 x 128. Reading the floats from the raw pointer is still on the CPU (also finding min and max).