Add gpu reading and saving
It is quite fast compared to CPU version, but there is some redundant operations because I have to keep converting from tex3d to rendtex and floats.
It would be a lot easier if tex3d would support some kind of decoding or "get floats" instead of colors which are not serializable.